requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
backoff==2.2.1
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
einops==0.8.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
gast==0.6.0
google-auth==2.41.1
google-genai==1.55.0
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kditransform==1.2.0
keras==2.15.0
kiwisolver==1.4.9
libclang==18.1.1
llvmlite==0.45.1
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mdatagen==0.2.0
missingno==0.5.2
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
openpyxl==3.1.5
openai==2.14.0
opt_einsum==3.4.0
packaging==25.0
pandas==2.2.2
pillow==12.0.0
platformdirs==4.5.0
posthog==6.9.0
protobuf==4.25.8
pyampute==0.0.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.16.3
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tabpfn==6.0.5
tabpfn-common-utils==0.2.7
-e git+https://github.com/priorlabs/tabpfn-extensions.git@43157b0c32d667be989d1a5b114092ceec3a899f#egg=tabpfn_extensions
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.2.0
threadpoolctl==3.6.0
torch==2.9.0
tqdm==4.66.1
triton==3.5.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.2
xlrd==2.0.2
